# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Note Taking App API",
    "version": "1.0.0",
    "description": "A comprehensive note-taking application with user authentication and note management features",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Format: Bearer <token>"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie authentication"
      }
    },
    "schemas": {
...
      "BlockDataSchema": {
        "oneOf": [
          {
            "type": "object"
          },
          {
            "type": "object"
          },
          {
            "type": "object"
          },
          {
            "type": "object"
          },
          {
            "type": "object"
          }
        ]
      },
      "EditorBlockSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique block identifier"
          },
          "type": {
            "type": "string",
            "enum": [
              "header",
              "paragraph",
              "list",
              "quote",
              "delimiter"
            ],
            "description": "Block type"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BlockDataSchema"
              }
            ],
            "description": "Block-specific data"
          }
        },
        "required": [
          "id",
          "type",
          "data"
        ]
      },
      "EditorContentSchema": {
        "type": "object",
        "properties": {
          "time": {
            "type": "number",
            "description": "Content creation/modification timestamp"
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EditorBlockSchema"
            },
            "description": "Array of content blocks"
          },
          "version": {
            "type": "string",
            "description": "Editor.js version"
          }
        },
        "required": [
          "time",
          "blocks",
          "version"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/notes/{id}": {
      "get": {
        "operationId": "get-notes-{id}",
        "summary": "Get note by ID",
        "description": "Retrieve a specific note by its ID with tags",
        "tags": [
          "Notes"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Note retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Response message"
                    },
                    "note": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "message",
                    "note",
                    "note"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "operationId": "put-notes-{id}",
        "summary": "Update note",
        "description": "Update an existing note's title, content, tags, or archive status",
        "tags": [
          "Notes"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "nullable": true,
                    "description": "Note title"
                  },
                  "content": {
                    "type": "object",
                    "description": "Note content in Editor.js format"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "nullable": true,
                    "description": "Array of tag names"
                  },
                  "isArchived": {
                    "type": "boolean",
                    "nullable": true,
                    "description": "Archive status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Response message"
                    },
                    "note": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "message",
                    "note",
                    "note"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-notes-{id}",
        "summary": "Delete note",
        "description": "Permanently delete a note and its associated tags",
        "tags": [
          "Notes"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Unique note identifier"
                    },
                    "userId": {
                      "type": "string",
                      "description": "Owner user ID"
                    },
                    "title": {
                      "type": "string",
                      "description": "Note title"
                    },
                    "content": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/EditorContentSchema"
                        }
                      ],
                      "description": "Note content in Editor.js format"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Array of tag names"
                    },
                    "isArchived": {
                      "type": "boolean",
                      "description": "Archive status"
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "Creation timestamp"
                    },
                    "updatedAt": {
                      "type": "string",
                      "description": "Last update timestamp"
                    },
                    "lastEdited": {
                      "type": "string",
                      "description": "Last edit timestamp"
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "title",
                    "content",
                    "tags",
                    "isArchived",
                    "createdAt",
                    "updatedAt",
                    "lastEdited"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}